I have completed the work with Fbx_Binary_File, now i can read a fbx file format and save the result in Fbx_Binary_File

+++ src/win32/win_types.h
@@ -6,7 +6,8 @@
 typedef BYTE    u8;
 typedef WORD    u16;
 typedef UINT    u32;
-typedef DWORD64 u64;
+typedef DWORD   u64;
+
 
 typedef INT8    s8;
 typedef INT16   s16;
